fix return type
git-svn-id: https://svn.salilab.org/imp/trunk@1910 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/misc/include/internal/tunnel_traits.h
@@ -59,7 +59,7 @@ class TunnelTraits {
   unsigned int get_coordinate() const {
     return coordinate_;
   }
-  unsigned int get_height() const {
+  Float get_height() const {
     return height_;
   }
   Float get_radius() const {
